Remove useless Tester\PHPUnit split from ParaTestCommand
