--- test/Unit/Console/Testers/TestCommand.php
@@ -1,15 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace ParaTest\Tests\Unit\Console\Testers;
-
-use Symfony\Component\Console\Command\Command;
-
-final class TestCommand extends Command
-{
-    public function __construct()
-    {
-        parent::__construct('testcommand');
-    }
-}
